+++ tests/Integration/Messaging/Controller/CampaignControllerTest.php
@@ -9,7 +9,6 @@ use PhpList\RestBundle\Tests\Integration\Common\AbstractTestController;
 use PhpList\RestBundle\Tests\Integration\Identity\Fixtures\AdministratorFixture;
 use PhpList\RestBundle\Tests\Integration\Identity\Fixtures\AdministratorTokenFixture;
 use PhpList\RestBundle\Tests\Integration\Messaging\Fixtures\MessageFixture;
-use PhpList\RestBundle\Tests\Integration\Subscription\Fixtures\SubscriberListFixture;
 
 class CampaignControllerTest extends AbstractTestController
 {
@@ -87,59 +86,5 @@ class CampaignControllerTest extends AbstractTestController
 
         $this->authenticatedJsonRequest('DELETE', '/api/v2/campaigns/1');
         $this->assertHttpNoContent();
-    }
-    public function testSendMessageWithoutSessionReturnsForbidden(): void
-    {
-        $this->loadFixtures([MessageFixture::class]);
-        self::getClient()->request('POST', '/api/v2/campaigns/1/send');
-        $this->assertHttpForbidden();
-    }
-
-    public function testSendMessageWithValidSessionReturnsOkay(): void
-    {
-        $this->loadFixtures([AdministratorFixture::class, MessageFixture::class]);
-
-        $this->authenticatedJsonRequest('POST', '/api/v2/campaigns/2/send');
-        $this->assertHttpOkay();
-
-        $response = $this->getDecodedJsonResponseContent();
-        self::assertSame(2, $response['id']);
-    }
-
-    public function testSendMessageWithInvalidIdReturnsNotFound(): void
-    {
-        $this->authenticatedJsonRequest('POST', '/api/v2/campaigns/999/send');
-        $this->assertHttpNotFound();
-    }
-
-    public function testResendMessageToListsWithoutSessionReturnsForbidden(): void
-    {
-        $this->loadFixtures([MessageFixture::class, SubscriberListFixture::class]);
-
-        $this->jsonRequest('POST', '/api/v2/campaigns/2/resend', [], [], [], json_encode(['list_ids' => [1]]));
-        $this->assertHttpForbidden();
-    }
-
-    public function testResendMessageToListsWithValidSessionReturnsOkay(): void
-    {
-        $this->loadFixtures([MessageFixture::class, SubscriberListFixture::class]);
-
-        $this->authenticatedJsonRequest('POST', '/api/v2/campaigns/2/resend', [], [], [], json_encode([
-            'list_ids' => [1],
-        ]));
-        $this->assertHttpOkay();
-
-        $response = $this->getDecodedJsonResponseContent();
-        self::assertSame(2, $response['id']);
-    }
-
-    public function testResendMessageToListsWithInvalidIdReturnsNotFound(): void
-    {
-        $this->loadFixtures([SubscriberListFixture::class]);
-
-        $this->authenticatedJsonRequest('POST', '/api/v2/campaigns/999/resend', [], [], [], json_encode([
-            'list_ids' => [1],
-        ]));
-        $this->assertHttpNotFound();
     }
 }
